move unit tests into src/ modules: numerics, tap_script_builder, block_filter_map, bls
