+++ core/indexer/src/bls.rs
@@ -720,4 +720,508 @@ mod tests {
             .expect_err("unresolvable signer_id must be rejected");
         assert!(resolver.pk_cache.is_empty());
     }
+
+    #[test]
+    fn rogue_key_forgery_succeeds_with_same_message() {
+        use crate::test_utils::bls_test::{construct_rogue_g2_pubkey, derive_test_key};
+        use blst::BLST_ERROR;
+
+        let victim_sk = derive_test_key(10);
+        let victim_pk = victim_sk.sk_to_pk();
+        let beta_sk = derive_test_key(11);
+        let beta_pk = beta_sk.sk_to_pk();
+        let rogue_pk_bytes = construct_rogue_g2_pubkey(&beta_pk.to_bytes(), &victim_pk.to_bytes());
+        let rogue_pk = blst::min_sig::PublicKey::key_validate(&rogue_pk_bytes)
+            .expect("rogue key must be valid");
+        let msg = b"identical-message";
+        let forged_sig = beta_sk.sign(msg, KONTOR_BLS_DST, &[]);
+        let result = forged_sig.aggregate_verify(
+            true,
+            &[msg.as_slice(), msg.as_slice()],
+            KONTOR_BLS_DST,
+            &[&rogue_pk, &victim_pk],
+            true,
+        );
+        assert_eq!(result, BLST_ERROR::BLST_SUCCESS);
+    }
+
+    #[test]
+    fn rogue_key_forgery_fails_with_distinct_messages() {
+        use crate::test_utils::bls_test::{construct_rogue_g2_pubkey, derive_test_key};
+        use blst::BLST_ERROR;
+        use blst::min_sig::AggregateSignature;
+
+        let victim_sk = derive_test_key(10);
+        let victim_pk = victim_sk.sk_to_pk();
+        let beta_sk = derive_test_key(11);
+        let beta_pk = beta_sk.sk_to_pk();
+        let rogue_pk_bytes = construct_rogue_g2_pubkey(&beta_pk.to_bytes(), &victim_pk.to_bytes());
+        let rogue_pk = blst::min_sig::PublicKey::key_validate(&rogue_pk_bytes)
+            .expect("rogue key must be valid");
+        let msg_attacker = b"attacker-op-data";
+        let msg_victim = b"victim-op-data";
+        let victim_sig = victim_sk.sign(msg_victim, KONTOR_BLS_DST, &[]);
+        let attacker_sig = beta_sk.sign(msg_attacker, KONTOR_BLS_DST, &[]);
+        let agg =
+            AggregateSignature::aggregate(&[&attacker_sig, &victim_sig], true).expect("aggregate");
+        let agg_sig = agg.to_signature();
+        let result = agg_sig.aggregate_verify(
+            true,
+            &[msg_attacker.as_slice(), msg_victim.as_slice()],
+            KONTOR_BLS_DST,
+            &[&rogue_pk, &victim_pk],
+            true,
+        );
+        assert_ne!(result, BLST_ERROR::BLST_SUCCESS);
+    }
+
+    #[test]
+    fn bls_attack_eve_registers_own_key_under_alice_identity_aggregate_rejected() {
+        use crate::test_utils::bls_test::derive_test_key;
+        use bitcoin::hashes::{Hash, sha256};
+        use bitcoin::key::{Keypair, Secp256k1, rand};
+        use bitcoin::secp256k1::Message;
+
+        let secp = Secp256k1::new();
+        let _alice_keypair = Keypair::new(&secp, &mut rand::thread_rng());
+        let eve_keypair = Keypair::new(&secp, &mut rand::thread_rng());
+        let alice_xonly = _alice_keypair.x_only_public_key().0;
+        let eve_bls_sk = derive_test_key(42);
+        let eve_bls_pk = eve_bls_sk.sk_to_pk();
+        let schnorr_msg = {
+            let mut preimage = Vec::with_capacity(SCHNORR_BINDING_PREFIX.len() + 96);
+            preimage.extend_from_slice(SCHNORR_BINDING_PREFIX);
+            preimage.extend_from_slice(&eve_bls_pk.to_bytes());
+            let digest = sha256::Hash::hash(&preimage).to_byte_array();
+            Message::from_digest_slice(&digest).expect("32-byte digest")
+        };
+        let eve_schnorr_sig = secp.sign_schnorr(&schnorr_msg, &eve_keypair).serialize();
+        let bls_binding_msg = {
+            let mut msg = Vec::with_capacity(BLS_BINDING_PREFIX.len() + 32);
+            msg.extend_from_slice(BLS_BINDING_PREFIX);
+            msg.extend_from_slice(&alice_xonly.serialize());
+            msg
+        };
+        let eve_bls_binding_sig = eve_bls_sk
+            .sign(&bls_binding_msg, KONTOR_BLS_DST, &[])
+            .to_bytes();
+        let op = Inst::RegisterBlsKey {
+            bls_pubkey: eve_bls_pk.to_bytes().to_vec(),
+            schnorr_sig: eve_schnorr_sig.to_vec(),
+            bls_sig: eve_bls_binding_sig.to_vec(),
+        };
+        let insts = Insts {
+            ops: vec![op],
+            aggregate: Some(AggregateInfo {
+                signer_ids: vec![0],
+                signature: vec![0u8; 48],
+            }),
+        };
+        let err = validate_aggregate_shape(&insts)
+            .expect_err("aggregate RegisterBlsKey must be rejected");
+        assert!(
+            err.to_string()
+                .contains("RegisterBlsKey is not allowed in aggregate")
+        );
+    }
+
+    fn call_op(
+        nonce: u64,
+        gas_limit: u64,
+        contract: ContractAddress,
+        expr: impl Into<String>,
+    ) -> Inst {
+        Inst::Call {
+            gas_limit,
+            contract,
+            nonce: Some(nonce),
+            expr: expr.into(),
+        }
+    }
+
+    #[test]
+    fn bls_bulk_aggregate_signature_roundtrip() {
+        use crate::test_utils::bls_test::derive_test_key;
+        use blst::BLST_ERROR;
+        let sk1 = derive_test_key(1);
+        let sk2 = derive_test_key(2);
+        let pk1 = sk1.sk_to_pk();
+        let pk2 = sk2.sk_to_pk();
+        let contract = ContractAddress {
+            name: "arith".into(),
+            height: 123,
+            tx_index: 4,
+        };
+        let op1 = call_op(0, 50_000, contract.clone(), "eval(10, id)");
+        let op2 = call_op(0, 50_000, contract, "eval(10, sum({y: 8}))");
+        let msg1 = op1.aggregate_signing_message(1).unwrap();
+        let msg2 = op2.aggregate_signing_message(2).unwrap();
+        let sig1 = sk1.sign(&msg1, KONTOR_BLS_DST, &[]);
+        let sig2 = sk2.sign(&msg2, KONTOR_BLS_DST, &[]);
+        let agg = blst::min_sig::AggregateSignature::aggregate(&[&sig1, &sig2], true).unwrap();
+        let msgs = [msg1, msg2];
+        let refs: Vec<&[u8]> = msgs.iter().map(Vec::as_slice).collect();
+        assert_eq!(
+            agg.to_signature()
+                .aggregate_verify(true, &refs, KONTOR_BLS_DST, &[&pk1, &pk2], true),
+            BLST_ERROR::BLST_SUCCESS
+        );
+    }
+
+    #[test]
+    fn bls_bulk_aggregate_signature_fails_if_op_bytes_change() {
+        use crate::test_utils::bls_test::derive_test_key;
+        use blst::BLST_ERROR;
+        let sk1 = derive_test_key(7);
+        let sk2 = derive_test_key(9);
+        let pk1 = sk1.sk_to_pk();
+        let pk2 = sk2.sk_to_pk();
+        let contract = ContractAddress {
+            name: "arith".into(),
+            height: 123,
+            tx_index: 4,
+        };
+        let msg1 = call_op(0, 50_000, contract.clone(), "eval(10, id)")
+            .aggregate_signing_message(1)
+            .unwrap();
+        let msg2 = call_op(0, 50_000, contract, "eval(10, sum({y: 8}))")
+            .aggregate_signing_message(2)
+            .unwrap();
+        let sig1 = sk1.sign(&msg1, KONTOR_BLS_DST, &[]);
+        let sig2 = sk2.sign(&msg2, KONTOR_BLS_DST, &[]);
+        let agg = blst::min_sig::AggregateSignature::aggregate(&[&sig1, &sig2], true).unwrap();
+        let msg1_mutated = call_op(
+            0,
+            60_000,
+            ContractAddress {
+                name: "arith".into(),
+                height: 123,
+                tx_index: 4,
+            },
+            "eval(10, id)",
+        )
+        .aggregate_signing_message(1)
+        .unwrap();
+        let msgs = [msg1_mutated, msg2];
+        let refs: Vec<&[u8]> = msgs.iter().map(Vec::as_slice).collect();
+        assert_ne!(
+            agg.to_signature()
+                .aggregate_verify(true, &refs, KONTOR_BLS_DST, &[&pk1, &pk2], true),
+            BLST_ERROR::BLST_SUCCESS
+        );
+    }
+
+    #[test]
+    fn bls_bulk_call_roundtrip_serialization_preserves_signer_id() {
+        let op = call_op(
+            7,
+            50_000,
+            ContractAddress {
+                name: "arith".into(),
+                height: 7,
+                tx_index: 3,
+            },
+            "eval(10, id)",
+        );
+        let bytes = indexer_types::serialize(&(42u64, op.clone())).unwrap();
+        let decoded: (u64, Inst) = indexer_types::deserialize(&bytes).unwrap();
+        assert_eq!(decoded, (42, op));
+    }
+
+    #[test]
+    fn bls_bulk_message_changes_when_signer_id_changes() {
+        let c = ContractAddress {
+            name: "arith".into(),
+            height: 123,
+            tx_index: 4,
+        };
+        assert_ne!(
+            call_op(0, 50_000, c.clone(), "eval(10, id)")
+                .aggregate_signing_message(1)
+                .unwrap(),
+            call_op(0, 50_000, c, "eval(10, id)")
+                .aggregate_signing_message(2)
+                .unwrap()
+        );
+    }
+
+    #[test]
+    fn bls_bulk_message_changes_when_nonce_changes() {
+        let c = ContractAddress {
+            name: "arith".into(),
+            height: 123,
+            tx_index: 4,
+        };
+        assert_ne!(
+            call_op(0, 50_000, c.clone(), "eval(10, id)")
+                .aggregate_signing_message(1)
+                .unwrap(),
+            call_op(1, 50_000, c, "eval(10, id)")
+                .aggregate_signing_message(1)
+                .unwrap()
+        );
+    }
+
+    #[test]
+    fn bls_bulk_message_changes_when_gas_limit_changes() {
+        let c = ContractAddress {
+            name: "arith".into(),
+            height: 123,
+            tx_index: 4,
+        };
+        assert_ne!(
+            call_op(0, 50_000, c.clone(), "eval(10, id)")
+                .aggregate_signing_message(1)
+                .unwrap(),
+            call_op(0, 60_000, c, "eval(10, id)")
+                .aggregate_signing_message(1)
+                .unwrap()
+        );
+    }
+
+    #[test]
+    fn bls_bulk_message_changes_when_contract_name_changes() {
+        assert_ne!(
+            call_op(
+                0,
+                50_000,
+                ContractAddress {
+                    name: "token".into(),
+                    height: 1,
+                    tx_index: 0
+                },
+                "transfer(\"x\", 10)"
+            )
+            .aggregate_signing_message(1)
+            .unwrap(),
+            call_op(
+                0,
+                50_000,
+                ContractAddress {
+                    name: "pool".into(),
+                    height: 1,
+                    tx_index: 0
+                },
+                "transfer(\"x\", 10)"
+            )
+            .aggregate_signing_message(1)
+            .unwrap()
+        );
+    }
+
+    #[test]
+    fn bls_bulk_message_changes_when_contract_height_changes() {
+        assert_ne!(
+            call_op(
+                0,
+                50_000,
+                ContractAddress {
+                    name: "token".into(),
+                    height: 1,
+                    tx_index: 0
+                },
+                "transfer(\"x\", 10)"
+            )
+            .aggregate_signing_message(1)
+            .unwrap(),
+            call_op(
+                0,
+                50_000,
+                ContractAddress {
+                    name: "token".into(),
+                    height: 2,
+                    tx_index: 0
+                },
+                "transfer(\"x\", 10)"
+            )
+            .aggregate_signing_message(1)
+            .unwrap()
+        );
+    }
+
+    #[test]
+    fn bls_bulk_message_changes_when_contract_tx_index_changes() {
+        assert_ne!(
+            call_op(
+                0,
+                50_000,
+                ContractAddress {
+                    name: "token".into(),
+                    height: 1,
+                    tx_index: 0
+                },
+                "transfer(\"x\", 10)"
+            )
+            .aggregate_signing_message(1)
+            .unwrap(),
+            call_op(
+                0,
+                50_000,
+                ContractAddress {
+                    name: "token".into(),
+                    height: 1,
+                    tx_index: 1
+                },
+                "transfer(\"x\", 10)"
+            )
+            .aggregate_signing_message(1)
+            .unwrap()
+        );
+    }
+
+    #[test]
+    fn bls_bulk_message_changes_when_expr_changes() {
+        let c = ContractAddress {
+            name: "token".into(),
+            height: 1,
+            tx_index: 0,
+        };
+        assert_ne!(
+            call_op(0, 50_000, c.clone(), "transfer(\"alice\", 10)")
+                .aggregate_signing_message(1)
+                .unwrap(),
+            call_op(0, 50_000, c, "transfer(\"bob\", 10)")
+                .aggregate_signing_message(1)
+                .unwrap()
+        );
+    }
+
+    #[test]
+    fn bls_bulk_wrong_signer_key_fails_single_op() {
+        use crate::test_utils::bls_test::derive_test_key;
+        use blst::BLST_ERROR;
+        let sk_a = derive_test_key(20);
+        let sk_b = derive_test_key(21);
+        let pk_a = sk_a.sk_to_pk();
+        let msg = call_op(
+            0,
+            50_000,
+            ContractAddress {
+                name: "token".into(),
+                height: 1,
+                tx_index: 0,
+            },
+            "transfer(\"dest\", 100)",
+        )
+        .aggregate_signing_message(1)
+        .unwrap();
+        let sig_by_b = sk_b.sign(&msg, KONTOR_BLS_DST, &[]);
+        assert_ne!(
+            sig_by_b.aggregate_verify(true, &[msg.as_slice()], KONTOR_BLS_DST, &[&pk_a], true),
+            BLST_ERROR::BLST_SUCCESS
+        );
+    }
+
+    #[test]
+    fn bls_bulk_wrong_signer_key_fails_multi_op_key_swap() {
+        use crate::test_utils::bls_test::derive_test_key;
+        use blst::BLST_ERROR;
+        let sk_a = derive_test_key(30);
+        let sk_b = derive_test_key(31);
+        let pk_a = sk_a.sk_to_pk();
+        let pk_b = sk_b.sk_to_pk();
+        let c = ContractAddress {
+            name: "token".into(),
+            height: 1,
+            tx_index: 0,
+        };
+        let msg_a = call_op(0, 50_000, c.clone(), "transfer(\"x\", 10)")
+            .aggregate_signing_message(1)
+            .unwrap();
+        let msg_b = call_op(0, 50_000, c, "transfer(\"y\", 20)")
+            .aggregate_signing_message(2)
+            .unwrap();
+        let agg = blst::min_sig::AggregateSignature::aggregate(
+            &[
+                &sk_b.sign(&msg_a, KONTOR_BLS_DST, &[]),
+                &sk_a.sign(&msg_b, KONTOR_BLS_DST, &[]),
+            ],
+            true,
+        )
+        .unwrap();
+        assert_ne!(
+            agg.to_signature().aggregate_verify(
+                true,
+                &[msg_a.as_slice(), msg_b.as_slice()],
+                KONTOR_BLS_DST,
+                &[&pk_a, &pk_b],
+                true
+            ),
+            BLST_ERROR::BLST_SUCCESS
+        );
+    }
+
+    #[test]
+    fn bls_bulk_one_correct_one_wrong_key_fails_entire_aggregate() {
+        use crate::test_utils::bls_test::derive_test_key;
+        use blst::BLST_ERROR;
+        let sk_a = derive_test_key(40);
+        let sk_b = derive_test_key(41);
+        let sk_c = derive_test_key(42);
+        let pk_a = sk_a.sk_to_pk();
+        let pk_b = sk_b.sk_to_pk();
+        let c = ContractAddress {
+            name: "token".into(),
+            height: 1,
+            tx_index: 0,
+        };
+        let msg_a = call_op(0, 50_000, c.clone(), "transfer(\"x\", 10)")
+            .aggregate_signing_message(1)
+            .unwrap();
+        let msg_b = call_op(0, 50_000, c, "transfer(\"y\", 20)")
+            .aggregate_signing_message(2)
+            .unwrap();
+        let agg = blst::min_sig::AggregateSignature::aggregate(
+            &[
+                &sk_a.sign(&msg_a, KONTOR_BLS_DST, &[]),
+                &sk_c.sign(&msg_b, KONTOR_BLS_DST, &[]),
+            ],
+            true,
+        )
+        .unwrap();
+        assert_ne!(
+            agg.to_signature().aggregate_verify(
+                true,
+                &[msg_a.as_slice(), msg_b.as_slice()],
+                KONTOR_BLS_DST,
+                &[&pk_a, &pk_b],
+                true
+            ),
+            BLST_ERROR::BLST_SUCCESS
+        );
+    }
+
+    mod proptest_bulk {
+        use super::*;
+        use proptest::prelude::*;
+
+        proptest! {
+            #[test]
+            fn signing_message_no_panic_on_arbitrary_call(
+                signer_id in any::<u64>(),
+                nonce in any::<u64>(),
+                gas_limit in any::<u64>(),
+                name in any::<String>(),
+                height in any::<u64>(),
+                tx_index in any::<u64>(),
+                expr in any::<String>(),
+            ) {
+                let op = call_op(nonce, gas_limit, ContractAddress { name, height, tx_index }, expr);
+                let msg = op.aggregate_signing_message(signer_id).expect("must not fail");
+                prop_assert!(!msg.is_empty());
+            }
+
+            #[test]
+            fn signing_message_no_panic_on_arbitrary_register(
+                signer_id in any::<u64>(),
+                bls_pubkey in proptest::collection::vec(any::<u8>(), 0..256),
+                schnorr_sig in proptest::collection::vec(any::<u8>(), 0..256),
+                bls_sig in proptest::collection::vec(any::<u8>(), 0..256),
+            ) {
+                let op = Inst::RegisterBlsKey { bls_pubkey, schnorr_sig, bls_sig };
+                let msg = op.aggregate_signing_message(signer_id).expect("must not fail");
+                prop_assert!(!msg.is_empty());
+            }
+        }
+    }
 }
